on_message_received/requirements.txt
flask
twilio
google-cloud-logging